Fixed bug where the wrong aggregator variable was being tested (#2)
* Fixed bug where the wrong aggregator variable was being tested

* Added a "UserData" parameter and equivalent accessor sub "userdata()" to
allow for the storage of arbitrary data with the peer. The main purpose of
this is to allow the user to store extra data (a scalar or ref) with the
peer that is then readable by the call back routines. I realise there are
other ways to do this, but this seems much cleaner.

* Fixes the situation where a socket is still in the list of sockets to
be selected on, yet it has been closed. I believe this is caused when
we create a new connection to a peer at the same time as we recieve one.
When we find ourselves with a bad fd, we recheck the list of sockets to
select on.

* It is possible to receive a notification message (error) in response
to an OPEN message (e.g. an unrecognised ASN). We were getting a
Finite State Machine error, now we call the notification callback

* _kill_session will call _close_session even if there socket is not open
this will finally terminate the session properly (stops some wierd loops).

* Added extra members of the notification state engine
Now calls _kill_session rather than _cease when the peer socket is closed